Rename the pages directory to views as it conflicts with Next's Page Router conventions and causes build time errors
