+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Home from "src/pages/Homepage";
+import Home from "src/views/Homepage";
 
 export default function HomePage() {
   return <Home />;
